# Adding benchmark for integrator and improving performance.