--- a/include/integrators/whittedIntegrator.h
+++ b/include/integrators/whittedIntegrator.h
@@ -8,7 +8,8 @@
 
 class WhittedIntegrator : public Integrator {
  private:
-  Scene* scene;
+  const IntersectableList* intersectableList;
+  const std::vector<PointLight*>* lights;
 
   // Check whether hitPosition receives light
   //
@@ -16,7 +17,7 @@ class WhittedIntegrator : public Integrator {
   // @param L light direction vector
   // @param t parameter of ray equation p_uvw(t) = 0 + t(s_uvw-0)
   // @return is light source occluded by object at hitPostion?
-  bool isOccluded(Vector3f* hitPosition, Vector3f* lightDir, float eps);
+  bool isOccluded(Vector3f* hitPosition, Vector3f* lightDir, float eps) const;
 
   // Compute BRDF contribution for a given source at closest intersection point.
   //
@@ -24,10 +25,11 @@ class WhittedIntegrator : public Integrator {
   // @param hitRecord closest intersection primary ray with scene.
   // @param t parameter of ray equation p_uvw(t) = 0 + t(s_uvw-0).
   // @return returns current spectrum of light source at intersaction point.
-  Spectrum* contributionOf(PointLight* lightSource, HitRecord* hitRecord);
+  Spectrum* contributionOf(const PointLight* lightSource, HitRecord* hitRecord) const;
 
  public:
-  WhittedIntegrator(Scene*);
+  WhittedIntegrator(const IntersectableList* intersectableList,
+                    const std::vector<PointLight*>* lights);
   Spectrum* integrate(const Ray&);
 };
 #endif