--- a/src/integrators/whittedIntegrator.cpp
+++ b/src/integrators/whittedIntegrator.cpp
@@ -23,8 +23,8 @@ bool WhittedIntegrator::isOccluded(Vector3f* hitPosition, Vector3f* lightDir, fl
   return hasShadowHit;
 }
 
-Spectrum* WhittedIntegrator::contributionOf(const PointLight* lightSource,
-                                            HitRecord* hitRecord) const {
+Spectrum WhittedIntegrator::contributionOf(const PointLight* lightSource,
+                                           HitRecord* hitRecord) const {
   HitRecord* lightHit = lightSource->sample();
   Vector3f lightDir(*lightHit->position);
   lightDir.sub(*hitRecord->position);
@@ -34,7 +34,7 @@ Spectrum* WhittedIntegrator::contributionOf(const PointLight* lightSource,
 
   if (isOccluded(hitRecord->position, &lightDir, d2)) {
     delete lightHit;
-    return new Spectrum();
+    return Spectrum(0);
   }
 
   auto brdfContribution = std::unique_ptr<Spectrum>(
@@ -59,20 +59,19 @@ Spectrum* WhittedIntegrator::contributionOf(const PointLight* lightSource,
   // find a better scaling approach
   contribution.scale(1.0 / d2);
 
-  return new Spectrum(contribution);
+  return Spectrum(contribution);
 }
 
 WhittedIntegrator::WhittedIntegrator(const IntersectableList* intersectableList,
                                      const std::vector<PointLight*>* lights)
     : intersectableList(intersectableList), lights(lights) {}
 
-Spectrum* WhittedIntegrator::integrate(const Ray& ray) {
+Spectrum WhittedIntegrator::integrate(const Ray& ray) const {
   int MAX_DEPTH = 5;
 
   auto hitRecord = std::unique_ptr<HitRecord>(intersectableList->intersect(ray));
   if (!hitRecord->isValid()) {
-
-    return new Spectrum();
+    return Spectrum();
   }
 
   Spectrum reflection;
@@ -83,10 +82,9 @@ Spectrum* WhittedIntegrator::integrate(const Ray& ray) {
     if (sample.isValid) {
       reflection.add(sample.brdf);
       Ray reflectedRay(new Vector3f(*hitRecord->position), new Vector3f(sample.w), ray.depth + 1);
-      Spectrum* spec = integrate(reflectedRay);
-      reflection.mult(*spec);
+      Spectrum spec = integrate(reflectedRay);
+      reflection.mult(spec);
       delete reflectedRay.origin;
-      delete spec;
     }
   }
 
@@ -97,28 +95,26 @@ Spectrum* WhittedIntegrator::integrate(const Ray& ray) {
 
       const Ray refractedRay(new Vector3f(*hitRecord->position), new Vector3f(sample.w),
                              ray.depth + 1);
-      Spectrum* spec = integrate(refractedRay);
-      refraction.mult(*spec);
+      Spectrum spec = integrate(refractedRay);
+      refraction.mult(spec);
 
       delete refractedRay.origin;
-      delete spec;
     }
   }
 
   if (hitRecord->material->hasSpecularReflection() ||
       hitRecord->material->hasSpecularRefraction()) {
-    Spectrum* total = new Spectrum();
-    total->add(reflection);
-    total->add(refraction);
+    Spectrum total = Spectrum(0);
+    total.add(reflection);
+    total.add(refraction);
 
     return total;
   }
 
-  Spectrum* contribution = new Spectrum();
+  Spectrum contribution = Spectrum(0);
   for (const PointLight* lightSource : *lights) {
-    Spectrum* currentContribution = contributionOf(lightSource, hitRecord.get());
-    contribution->add(*currentContribution);
-    delete currentContribution;
+    Spectrum currentContribution = contributionOf(lightSource, hitRecord.get());
+    contribution.add(currentContribution);
   }
 
   return contribution;