Benchmarks indicate 5% better performance:
```
RUNNING: ./benchmarks_int_base.x --benchmark_repetitions=10 --benchmark_filter=BM_WhittedIntegrator.* --benchmark_display_aggregates_only=true --benchmark_out=/tmp/tmpfwhQun
2019-03-17 15:45:22
Running ./benchmarks_int_base.x
Run on (4 X 3000 MHz CPU s)
CPU Caches:
  L1 Data 32K (x2)
  L1 Instruction 32K (x2)
  L2 Unified 256K (x2)
  L3 Unified 4096K (x1)
Load Average: 0.29, 0.63, 0.81
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
BM_WhittedIntegrator_mean         6269 ns         6268 ns           10
BM_WhittedIntegrator_median       6273 ns         6273 ns           10
BM_WhittedIntegrator_stddev       15.6 ns         15.4 ns           10
RUNNING: ./benchmarks.x --benchmark_repetitions=10 --benchmark_filter=BM_WhittedIntegrator.* --benchmark_display_aggregates_only=true --benchmark_out=/tmp/tmpVicew6
2019-03-17 15:45:29
Running ./benchmarks.x
Run on (4 X 3000 MHz CPU s)
CPU Caches:
  L1 Data 32K (x2)
  L1 Instruction 32K (x2)
  L2 Unified 256K (x2)
  L3 Unified 4096K (x1)
Load Average: 0.35, 0.64, 0.81
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
BM_WhittedIntegrator_mean         5973 ns         5972 ns           10
BM_WhittedIntegrator_median       5968 ns         5968 ns           10
BM_WhittedIntegrator_stddev       20.0 ns         20.1 ns           10
Comparing ./benchmarks_int_base.x to ./benchmarks.x
Benchmark                                     Time             CPU      Time Old      Time New       CPU Old       CPU New
--------------------------------------------------------------------------------------------------------------------------
BM_WhittedIntegrator_pvalue                 0.0002          0.0002      U Test, Repetitions: 10 vs 10
BM_WhittedIntegrator_mean                  -0.0472         -0.0472          6269          5973          6268          5972
BM_WhittedIntegrator_median                -0.0486         -0.0487          6273          5968          6273          5968
BM_WhittedIntegrator_stddev                +0.2842         +0.3028            16            20            15            20
```